only delete working dir on cleanup when not in CI

+++ scripts/build.sh
@@ -153,7 +153,9 @@ cleanup() {
             done
         fi
         popd &> /dev/null
-        rm -rf $(dirname "$WORKING_DIRECTORY")
+        if [[ "$WORKING_DIRECTORY" =~ 'tempci' ]]; then
+            rm -rf $(dirname "$WORKING_DIRECTORY")
+        fi
     else
         echo "Workspace dir: $WORKSPACE"
         echo "Working Dir: $WORKING_DIRECTORY"
